requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
httpx==0.28.1
pydantic-settings==2.7.0
